requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
pyyaml>=6.0
onnx>=1.14.0
onnxruntime>=1.15.0
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
Pillow>=10.0.0
